+++ qiskit/assembler/assemble_schedules.py
@@ -25,7 +25,7 @@ from qiskit.qobj.converters import InstructionToQobjConverter, LoConfigConverter
 logger = logging.getLogger(__name__)
 
 
-def assemble_schedules(schedules, qobj_id=None, qobj_header=None, run_config=None):
+def assemble_schedules(schedules, qobj_id, qobj_header, run_config):
     """Assembles a list of schedules into a qobj which can be run on the backend.
     Args:
         schedules (list[Schedule]): schedules to assemble
@@ -83,7 +83,7 @@ def assemble_schedules(schedules, qobj_id=None, qobj_header=None, run_config=Non
     qobj_config['pulse_library'] = [PulseLibraryItem(name=pulse.name, samples=pulse.samples)
                                     for pulse in user_pulselib]
 
-    # create qob experiment field
+    # create qobj experiment field
     experiments = []
     schedule_los = qobj_config.pop('schedule_los', [])