+++ qiskit/quantum_info/operators/measures.py
@@ -12,7 +12,7 @@
 # copyright notice, and modified files need to carry a notice indicating
 # that they have been altered from the originals.
 
-# pylint: disable=invalid-name,anomalous-backslash-in-string
+# pylint: disable=invalid-name
 
 """
 A collection of useful quantum information functions for operators.
@@ -20,7 +20,7 @@ A collection of useful quantum information functions for operators.
 
 import numpy as np
 
-from qiskit.qiskiterror import QiskitError
+from qiskit.exceptions import QiskitError
 from qiskit.quantum_info.operators import Operator
 from qiskit.quantum_info.operators import SuperOp